New package scanner: pushing files' SHA256 to a webservice.
Because the RPM- or Deb-databases might be wrong (by accident,
or if someone "helpfully" copies a newer or older library in),
this scanner calculates the SHA256 of all files in a layer and
pushes this list to a web-service to resolve it into a list of
package names and versions.

This also works for minimized images - even better, in fact.